new icons and move some things
