+++ Optics/UINavigationExtensions.swift
@@ -1,0 +1,22 @@
+//
+//  UINavigationExtensions.swift
+//  Optics
+//
+//  Created by Jérémy Smith on 25/03/2016.
+//  Copyright © 2016 Jérémy Smith. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+// To custom status bar to light content
+extension UINavigationController {
+    
+    public override func childViewControllerForStatusBarHidden() -> UIViewController? {
+        return self.topViewController
+    }
+    
+    public override func childViewControllerForStatusBarStyle() -> UIViewController? {
+        return self.topViewController
+    }
+}
